requirements.txt
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
pyperclip==1.9.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
pyzbar==0.1.9
